+++ packages/react/src/twilio/components/TwilioWaitingPage.tsx
@@ -15,7 +15,7 @@ interface TwilioWaitingPageProps {
 
 type Step = "setup" | "connect" | "join" | "ready";
 
-export default function TwilioWaitingPage(props: TwilioWaitingPageProps) {
+function TwilioWaitingPage(props: TwilioWaitingPageProps) {
   const { spaceId, roomName, onRoomChanged, blameDictionary } = props;
   const [step, setStep] = useState<Step>("setup");
   const jwtToken = useJwtToken();
@@ -28,7 +28,11 @@ export default function TwilioWaitingPage(props: TwilioWaitingPageProps) {
     audioError,
   } = useTwilioLocalTracks();
   const [changeRoom, setChangeRoom] = useState(false);
-  const { join, isFull } = useJoin(spaceId, changeRoom, props.roomName);
+  const { join, isFull, isJoining } = useJoin(
+    spaceId,
+    changeRoom,
+    props.roomName,
+  );
   const onJoinClicked = useCallback((change: boolean) => {
     setChangeRoom(change);
     setStep("connect");
@@ -118,6 +122,7 @@ export default function TwilioWaitingPage(props: TwilioWaitingPageProps) {
       disabled={step !== "setup"}
       isAcquiringLocalTracks={isAcquiringLocalTracks}
       roomFull={isFull}
+      isJoining={isJoining || step !== "setup"}
       audioError={audioError}
       videoError={videoError}
       blameDictionary={blameDictionary}
@@ -128,3 +133,5 @@ export default function TwilioWaitingPage(props: TwilioWaitingPageProps) {
 TwilioWaitingPage.defaultProps = {
   roomName: null,
 };
+
+export default React.memo<TwilioWaitingPageProps>(TwilioWaitingPage);